requirements.txt
python-musicpd
wxpython